requirements.txt
flask
flask-session
spotipy
python-dotenv